Refactor: 채팅 scroll 관련 hook 리팩토링 (MM-55)
