+++ apps/mobile/app/index.tsx
@@ -21,6 +21,7 @@ export default function App() {
   const { setKeyboardHeight } = useBridge(appBridge)
   const insets = useSafeAreaInsets()
 
+  // const webviewUrl = process.env.EXPO_PUBLIC_WEB_VIEW_URL
   const webviewUrl =
     Platform.OS === 'android' ? process.env.EXPO_PUBLIC_ANDROID_WEB_VIEW_URL : process.env.EXPO_PUBLIC_IOS_WEB_VIEW_URL
